test: sort and clean snapshots
