+++ internal/output/testmain_test.go
@@ -1,0 +1,16 @@
+package output_test
+
+import (
+	"os"
+	"testing"
+
+	"github.com/google/osv-scanner/internal/testsnapshot"
+)
+
+func TestMain(m *testing.M) {
+	code := m.Run()
+
+	testsnapshot.Clean(m)
+
+	os.Exit(code)
+}
